+++ tests/noise_model/EmulatorNoiseModelTester.cpp
@@ -28,4 +28,13 @@ TEST(EmulatorNoiseModelTester, check_nm2)
   EXPECT_EQ(qubits.size(), 64);
 }
 
-
+TEST(EmulatorNoiseModelTester, check_nm3)
+{
+  qb::NoiseModel noise_model("qb-nm3");
+  std::set<int> qubits;
+  for (const auto& [q1, q2]: noise_model.get_connectivity()) {
+    qubits.emplace(q1);
+    qubits.emplace(q2);
+  }
+  EXPECT_EQ(qubits.size(), 48);
+}